Add files via upload
----------
TASK 1
----------
1.Print the square of all numbers from LOW to HIGH. Take low and high value from user in
input.
2. Take a number as an input from the user. Find if the number is even or odd.
3. Write a program that reads in five integers from user and determines and prints the largest
and the smallest integers in the group using relational operators.
4. Write a program that reads in two integers and determines and prints if the first is a multiple
of the second. (Hint: Use the modulus operator. For multiple, one number must be divisible to
the other number.). You might need to use simple if else statement to print if the number is a
multiple or not.
5. Print the following pattern. You must use loops. The number of loops must not be not exceed
2!
----------
TASK 2
----------
Given the immutability of strings do the following:
1) Create a string by taking input from the user (take at least 10 length input). Use “.” as a
delimiting character to stop taking input.
2) Take another substring subs as an input. Find and replace the contents of a string from
index 4 to 9 (inclusive) with three times the substring subs. You must use slicing and
single statement to fulfil that.
3) Change the string such that subs is appended between index 3 and 4. You might need to
use slicing here as you cannot update the index of string because of its immutability.
----------
TASK 3
----------
1) Take four inputs from the user. You must use ONLY a SINLGE line with input function to
take comma separated inputs. Use split function and make sure that the program does
not crashes if the user unknowingly enters more than four values. The values MUST be
stored in a list.
2) Parse the list, extract four values, and then type cast these values to integers.3) Prepare two fractions F1 and F2 using Fraction module.
4) Perform the following operations and print the results also
a. Add (F1+F2)
b. Subtract (F2-F1)
----------
TASK 4
----------
1) Take space separated 10 values from the user (in a list). Assume that the user will
always enter 10 single digit integers.
2) Using a single statement and comprehension, triple each value and concatenate it to the
existing list.
----------
TASK 5
----------
Write a program that
1) Creates a new text file and fill its contents with the data as:
a. Write 10 lines. Each line of a file must contain two entries separated by comma (,)
b. The two entries must be generated randomly. The first entry must be a random
number between 0 and 1. The second entry must be a random number lying in a
range [10-50]
2) Read the file created in step 1 and prepare a dictionary with all the file contents. Display
that dictionary too.
3) Take two inputs from the user; key (k) and value to update (val). Update the dictionary
structure by changing the value associated with k to val. Print the updated structure too.